+++ src/com/example/lansongeditordemo/VViewCommonWidgetActivity.java
@@ -9,15 +9,16 @@ import java.util.TimerTask;
 import java.util.concurrent.atomic.AtomicInteger;
 
 
-import net.frakbot.jumpingbeans.JumpingBeans;
 
 import jp.co.cyberagent.lansongsdk.gpuimage.GPUImageFilter;
 
-import com.daimajia.numberprogressbar.NumberProgressBar;
-import com.daimajia.numberprogressbar.OnProgressBarListener;
+import com.example.lansong.animview.JumpingBeans;
+import com.example.lansong.animview.MatchTextView;
+import com.example.lansong.animview.NumberProgressBar;
+import com.example.lansong.animview.OnProgressBarListener;
+import com.example.lansong.animview.PaintConstants;
 import com.example.lansongeditordemo.view.GLRelativeLayout;
 import com.example.lansongeditordemo.view.MediaPoolView;
-import com.jiangjie.utils.PaintConstants;
 import com.lansoeditor.demo.R;
 import com.lansosdk.box.ViewSprite;
 import com.lansosdk.box.ISprite;
@@ -26,7 +27,6 @@ import com.lansosdk.videoeditor.MediaInfo;
 import com.lansosdk.videoeditor.SDKDir;
 import com.lansosdk.videoeditor.SDKFileUtils;
 import com.lansosdk.videoeditor.VideoEditor;
-import com.roger.match.library.MatchTextView;
 
 import android.app.ActionBar.LayoutParams;
 import android.app.Activity;
@@ -222,8 +222,13 @@ public class VViewCommonWidgetActivity extends Activity implements OnSeekBarChan
 						toastStop();
 						
 						if(SDKFileUtils.fileExist(editTmpPath)){
-							VideoEditor.encoderAddAudio(mVideoPath,editTmpPath,SDKDir.TMP_DIR, dstPath);
-							SDKFileUtils.deleteFile(editTmpPath);
+							boolean ret=VideoEditor.encoderAddAudio(mVideoPath,editTmpPath,SDKDir.TMP_DIR, dstPath);
+							if(!ret){
+								dstPath=editTmpPath;
+							}else{
+								SDKFileUtils.deleteFile(editTmpPath);	
+							}
+							
 					    	findViewById(R.id.id_vview_realtime_saveplay).setVisibility(View.VISIBLE);
 						}
 					}
