+++ src/com/example/lansongeditordemo/VideoRemarkActivity.java
@@ -147,8 +147,12 @@ public class VideoRemarkActivity extends Activity{
 						toastStop();
 						
 						if(SDKFileUtils.fileExist(editTmpPath)){
-							VideoEditor.encoderAddAudio(mVideoPath,editTmpPath,SDKDir.TMP_DIR,dstPath);
-							SDKFileUtils.deleteFile(editTmpPath);
+							boolean ret=VideoEditor.encoderAddAudio(mVideoPath,editTmpPath,SDKDir.TMP_DIR,dstPath);
+							if(!ret){
+								dstPath=editTmpPath;
+							}else
+								SDKFileUtils.deleteFile(editTmpPath);
+							
 							findViewById(R.id.id_mediapool_saveplay).setVisibility(View.VISIBLE);
 						}
 					}
